[TASK] Show message if no extensions are installed
If EXT:extension_compatibility_check is the only extension installed in
the system, the list (table) does not appear empty any more. A message
points out that no extensions are currently installed.

In addition, buttons such as "execute check" make no sense in this case
and are removed from the view.

Resolves: #9